api/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
openpyxl